Implement implicit convertions.
Conversions were based on how C does it, but it's open for changes
Conversions are applied when a certain type of expression is required.
This includes:
- Unary operators
- Binary operators (operand's types are used to deduce a common type)
- Operations that require rvalues
- Variable initializers & assignments
- Operations where a boolean is expected (if's and while's)

Conversions currently not supported:
- Function arguments (this requires function type information at call site, currently only stores return type)
- Pointer arithmetic
- Conversions between different pointer types (such as constness cast)

+++ src/parser/syntax.cpp
@@ -57,7 +57,7 @@ nullptr_literal parser::build_null_ptr_lit(token_location loc)
 {
     nullptr_literal node;
     node.type.basetp = dtypename::_none;
-    node.type.exttp.push_back(static_cast<std::byte>(dtypemod::_ptr));
+    node.type.exttp.push_back(dtypemod::_ptr);
     node.cat = exprcat::rval;
     node.loc = loc;
     return node;
